Use elegant solution from vanished answer
There was a second answer to the question
http://stackoverflow.com/questions/27771508/showing-markdown-content-in-a-div
which suddenly disappeared mysteriously.

The answer described an elegant solution without the need to capture the
contents of 'about-me.md'. I tried to implement it, but wasn't able to
get it to work.

The solution essentially involves making a separate about-me layout
which contains the two divs. Parameters to the divs are passed using
frontmatter variables. The problem I faced was not being able to pass
the about-me text as the content.